refactor(cycle-section): components folder reorder
