+++ src/components/cycle-section/learning-units-section/complete-learning-unit-toogle/CompleteLearningUnitToogle.module.scss
