+++ src/components/cycle-section/learning-units-section/learning-unit-list-item/LearningUnitListItem.module.scss
